+++ RMK_support/rk_wrapper.py
@@ -110,6 +110,7 @@ class RKWrapper:
                 "load": False,  # If true, restart data will be loaded at the start of the loop. Will throw error if loadInitValsFromHDF5 is also true
                 "frequency": 1,  # Restart save frequency - saving every n steps
                 "resetTime": False,  # If true, the value of the time variable, if present, will be reset to 0 on loading from restart
+                "initialOutputIndex": 0,  # Initial output index, useful when restarting so that no data is overwritten
             },
             "loadInitValsFromHDF5": False,  # True if variables should be loaded from a complete HDF5 file based on the input vars list in the HDF5 options.
             "initValFilename": "ReMKiT1DVarInput",  # Name of the input hdf5 file
@@ -823,7 +824,7 @@ class RKWrapper:
         self.__timeloopData__["minimumSaveInterval"] = minimumInterval
 
     def setRestartOptions(
-        self, save=False, load=False, frequency=1, resetTime=False
+        self, save=False, load=False, frequency=1, resetTime=False, initialOutputIndex=0
     ) -> None:
         """Set restart options in timeloop object
 
@@ -832,6 +833,7 @@ class RKWrapper:
             load (bool, optional): Set to true if the code should initialize from restart data. Defaults to False.
             frequency (int, optional): Frequency at which restart data is saved in timesteps. Defaults to 1.
             resetTime (bool, optional): Set to true if the code should reset the time variable on restart. Defaults to False.
+            initialOutputIndex (int, optional): Sets the first output file index. Useful when avoiding overwriting files. Defaults to 0, outputting the initial value and grid.
         """
 
         cast(Dict[str, object], self.__timeloopData__["restart"])["save"] = save
@@ -842,6 +844,9 @@ class RKWrapper:
         cast(Dict[str, object], self.__timeloopData__["restart"])[
             "resetTime"
         ] = resetTime
+        cast(Dict[str, object], self.__timeloopData__["restart"])[
+            "initialOutputIndex"
+        ] = initialOutputIndex
 
     def setHDF5FileInitialData(
         self,